# DF/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/usertransactions/schema": {
      "get": {
        "description": "Retrive the datatype of the usertransactions table",
        "operationId": "usertransactionsController_findSchema",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usertransactionsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/usertransactions/get": {
      "get": {
        "description": "Filter all the records from the usertransactions table",
        "operationId": "usertransactionsController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/usertransactionsEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/usertransactions/{transactionid}": {
      "get": {
        "description": "Read only one records from the usertransactions table",
        "operationId": "usertransactionsController_findOne",
        "parameters": [
          {
            "name": "transactionid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usertransactionsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the only one record",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the usertransactions table",
        "operationId": "usertransactionsController_update",
        "parameters": [
          {
            "name": "transactionid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateusertransactionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usertransactionsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the usertransactions table",
        "operationId": "usertransactionsController_remove",
        "parameters": [
          {
            "name": "transactionid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usertransactionsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/usertransactions": {
      "get": {
        "description": "Read all the records from the usertransactions table",
        "operationId": "usertransactionsController_findAll",
        "parameters": [
          {
            "name": "trs_creator_email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_created_date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "trs_created_by",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_modified_date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "trs_modified_by",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_next_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_process_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_access_profile",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_org_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_org_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_role_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_role_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_ps_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_ps_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_sub_org_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_sub_org_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/usertransactionsEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the usertransactions table",
        "operationId": "usertransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateusertransactionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usertransactionsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/usertransactions/find/first": {
      "get": {
        "description": "Read first record from the usertransactions table",
        "operationId": "usertransactionsController_findFirst",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usertransactionsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the first record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/usertransactions/find/last": {
      "get": {
        "description": "Read last record from the usertransactions table",
        "operationId": "usertransactionsController_findLast",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/usertransactionsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the last record",
        "tags": [
          "ERD API"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "The Nestjs API documentation",
    "version": "0.4",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://192.168.2.57:3010",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "usertransactionsEntity": {
        "type": "object",
        "properties": {
          "transactionid": {
            "type": "number",
            "example": "number"
          },
          "transactiondate": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "amount": {
            "type": "number",
            "example": "number"
          },
          "transactiontype": {
            "type": "string",
            "example": "string"
          },
          "description": {
            "type": "string",
            "example": "string"
          },
          "name": {
            "type": "string",
            "example": "string"
          },
          "trs_creator_email": {
            "type": "string",
            "example": "string"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_created_by": {
            "type": "string",
            "example": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_modified_by": {
            "type": "string",
            "example": "string"
          },
          "trs_status": {
            "type": "string",
            "example": "string"
          },
          "trs_next_status": {
            "type": "string",
            "example": "string"
          },
          "trs_process_id": {
            "type": "string",
            "example": "string"
          },
          "trs_access_profile": {
            "type": "string",
            "example": "string"
          },
          "trs_org_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "transactionid",
          "transactiondate",
          "amount",
          "transactiontype",
          "description",
          "name",
          "trs_creator_email",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code",
          "trs_sub_org_code",
          "trs_sub_org_grp_code"
        ]
      },
      "CreateusertransactionsDto": {
        "type": "object",
        "properties": {
          "transactiondate": {
            "type": "string",
            "format": "date-time"
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "transactiontype": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "trs_creator_email": {
            "type": "string"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string"
          },
          "trs_sub_org_code": {
            "type": "string"
          }
        },
        "required": [
          "transactiondate",
          "amount",
          "transactiontype",
          "description",
          "name",
          "trs_creator_email",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code",
          "trs_sub_org_grp_code",
          "trs_sub_org_code"
        ]
      },
      "UpdateusertransactionsDto": {
        "type": "object",
        "properties": {
          "transactiondate": {
            "type": "string",
            "format": "date-time"
          },
          "amount": {
            "type": "integer",
            "format": "int32"
          },
          "transactiontype": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "trs_creator_email": {
            "type": "string",
            "example": "string"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          },
          "trs_sub_org_code": {
            "type": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string"
          }
        },
        "required": [
          "transactiondate",
          "amount",
          "transactiontype",
          "description",
          "name",
          "trs_creator_email",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code",
          "trs_sub_org_code",
          "trs_sub_org_grp_code"
        ]
      },
...
    }
  }
}